ClientApp.Tests/UsersControllerTests.cs
using System.Linq;
using Xunit;
using System.Collections.Generic;
using System.Threading.Tasks;
using Moq;
using Application.Interfaces;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Logging;
using AutoMapper;
using Microsoft.Extensions.DependencyInjection;
using ClientApp.Controllers;
using Domain.Entities.Documents;
using Application.Services.Communication;
using ClientApp.Controllers.Resources;

using ClientApp.Mapping;
using Application.Security.Services;
using Domain.Entities.Security;
using Application.Security.Services.Communication;
using System;
using ClientApp.Controllers.Resources.Security;

namespace ClientApp.Tests
{
    public class UsersControllerTests
    {
        IMapper mapper;
        Mock<IUserService> mockUserService;

        //This test is not required as it will always fails because the model will always be true because the model will not work in testing mode
        //[Fact]
        //public async Task CreateUserAsync_ReturnsBadRequest_WhenModelStateIsNotValid()
        //{
        //    //Arrange  
        //    CreateMockedObjects();
        //    //mockUserService.Setup(s => s.CreateUserAsync(It.IsAny<User>(), It.IsAny<ERole>()))
        //    //        .ReturnsAsync(GetFakeSucceededCreateUserResponse());
        //    var controller = new UsersController(mockUserService.Object, mapper);

        //    //Act
        //    var result = await controller.CreateUserAsync(GetFakeInvalidUserCredentialsResource());

        //    //Assert
        //    var badRequestResult = Assert.IsType<BadRequestObjectResult>(result);
        //    Assert.Equal("Failed to create user!", badRequestResult.Value);
        //}

        [Fact]
        public async Task CreateUserAsync_ReturnsBadRequest_WhenFailedToCreateTheUser()
        {
            //Arrange  
            CreateMockedObjects();
            mockUserService.Setup(s => s.CreateUserAsync(It.IsAny<User>(), It.IsAny<ERole>()))
                    .ReturnsAsync(GetFakeFailedCreateUserResponse());
            var controller = new UsersController(mockUserService.Object, mapper);

            //Act
            var result = await controller.CreateUserAsync(GetFakeValidUserCredentialsResource());

            //Assert
            var badRequestResult = Assert.IsType<BadRequestObjectResult>(result);
            Assert.Equal("Failed to create user!", badRequestResult.Value);
        }

        [Fact]
        public async Task CreateUserAsync_ReturnsOkResultWithTheUserResource_WhenSucceededToCreateTheUser()
        {
            //Arrange  
            CreateMockedObjects();
            mockUserService.Setup(s => s.CreateUserAsync(It.IsAny<User>(), It.IsAny<ERole>()))
                    .ReturnsAsync(GetFakeSucceededCreateUserResponse());
            var controller = new UsersController(mockUserService.Object, mapper);

            //Act
            var result = await controller.CreateUserAsync(GetFakeValidUserCredentialsResource());

            //Assert
            var okResult = Assert.IsType<OkObjectResult>(result);
            var userResource = Assert.IsAssignableFrom<UserResource>(okResult.Value);
            Assert.Equal(1, userResource.Id);
            Assert.Equal("test", userResource.Email);
        }

        private UserCredentialsResource GetFakeInvalidUserCredentialsResource()
        {
            return new UserCredentialsResource() { Email="a" };
        }

        private UserCredentialsResource GetFakeValidUserCredentialsResource()
        {
            return new UserCredentialsResource
            {
                Email = "test",
                Password = "test"
            };
        }

        private CreateUserResponse GetFakeFailedCreateUserResponse()
        {
            return new CreateUserResponse(false, "Failed to create user!", null);            
        }

        private CreateUserResponse GetFakeSucceededCreateUserResponse()
        {
            return new CreateUserResponse(true, "", new User() { Id = 1, Email = "test", Password = "test" });
        }

        private void CreateMockedObjects()
        {
            mapper = CreateMapper();
            mockUserService = new Mock<IUserService>();
        }
        private IMapper CreateMapper()
        {
            //var mockIMapper = new Mock<IMapper>();
            var config = new MapperConfiguration(opts =>
            {
                // Add your mapper profile configs or mappings here
                opts.AddProfile<ModelToResourceProfile>();
                opts.AddProfile<ResourceToModelProfile>();
            });
            var mapper = config.CreateMapper();
            return mapper;
        }
    }
}
